Fix Add Required Parameter
In file: blueprint.py, class: BlueprintMixin, method: <method_name> does not follow the naming convention for its first parameter. This may be an instance method with the first parameter not called self, or a class method with the first parameter not called cls. iCR identified these and suggested that the parameter name is changed to follow convention.